refactor: rename invitation to invite command
