+++ src/Skak.Discord/Utils/DiscordClientExtensions.cs
@@ -23,7 +23,7 @@ namespace Skak.Discord.Utils
             // Add new command modules here
             slash.RegisterCommands<HealthCommands>();
             slash.RegisterCommands<PodiumCommands>();
-            slash.RegisterCommands<InvitationCommands>();
+            slash.RegisterCommands<InviteCommand>();
             slash.RegisterCommands<RankingCommand>();
 
             return slash;
